+++ src/js/common/schemaform/save-in-progress/SaveInProgressErrorPage.jsx
@@ -14,7 +14,7 @@ import {
 import SignInLink from '../../../../platform/forms/components/SignInLink';
 import ProgressButton from '@department-of-veterans-affairs/jean-pants/ProgressButton';
 
-import { toggleLoginModal } from '../../../../platform/user/login/actions';
+import { toggleLoginModal } from '../../../../platform/site-wide/user-nav/actions';
 
 
 // For now, this only handles loading errors, but it could feasibly be reworked
