+++ src/js/disability-benefits/526EZ/components/IntroductionPage.jsx
@@ -8,7 +8,7 @@ import { focusElement } from '../../../../platform/utilities/ui';
 import OMBInfo from '@department-of-veterans-affairs/jean-pants/OMBInfo';
 import FormTitle from '../../../common/schemaform/components/FormTitle';
 import { introActions, introSelector } from '../../../common/schemaform/save-in-progress/SaveInProgressIntro';
-import { toggleLoginModal } from '../../../../platform/user/login/actions';
+import { toggleLoginModal } from '../../../../platform/site-wide/user-nav/actions';
 
 import formConfig from '../config/form';
 import { submitIntentToFile } from '../actions';
