+++ src/js/disability-benefits/526EZ/components/VerifiedReviewContainer.jsx
@@ -12,7 +12,7 @@ import {
   saveAndRedirectToReturnUrl,
   autoSaveForm
 } from '../../../common/schemaform/save-in-progress/actions';
-import { toggleLoginModal } from '../../../../platform/user/login/actions';
+import { toggleLoginModal } from '../../../../platform/site-wide/user-nav/actions';
 import {
   getNextPagePath,
   getPreviousPagePath
