+++ src/platform/site-wide/user-nav/components/SignInProfileMenu.jsx
@@ -4,7 +4,7 @@ import React from 'react';
 import DropDownPanel from '@department-of-veterans-affairs/jean-pants/DropDownPanel';
 import IconUser from '@department-of-veterans-affairs/jean-pants/IconUser';
 import { features } from '../../../../js/personalization/beta-enrollment/containers/BetaApp';
-import { logout } from '../../../user/login/utilities';
+import { logout } from '../../../user/authentication/utilities';
 import BetaDropdown from './BetaDropdown';
 
 class SignInProfileMenu extends React.Component {
