+++ src/platform/site-wide/user-nav/reducers/index.js
