+++ src/platform/user/authentication/utilities.js
