+++ src/platform/user/profile/actions/index.js
@@ -1,7 +1,7 @@
 import recordEvent from '../../../monitoring/record-event';
 import { removeFormApi } from '../../../../js/common/schemaform/save-in-progress/api';
 import { apiRequest } from '../../../utilities/api';
-import { updateLoggedInStatus } from '../../../user/login/actions';
+import { updateLoggedInStatus } from '../../../site-wide/user-nav/actions';
 import environment from '../../../utilities/environment';
 
 export const UPDATE_PROFILE_FIELDS = 'UPDATE_PROFILE_FIELDS';
