+++ test/common/schemaform/save-in-progress/actions.unit.spec.js
@@ -20,7 +20,7 @@ import {
   setStartOver,
 } from '../../../../src/js/common/schemaform/save-in-progress/actions';
 
-import { logOut } from '../../../../src/platform/user/login/actions';
+import { logOut } from '../../../../src/platform/site-wide/user-nav/actions';
 
 let oldFetch;
 let oldSessionStorage;
